converted use of floats to ints, created the logic for the edit goal fragment, attached the edit fragment to day, week, month instance.

+++ app/src/main/java/com/example/mwt/db/dateprogressdb/DateProgressEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "dateProgress")
 data class DateProgressEntity(
         val date: String,
-        val progress: Float
+        val progress: Int
 ) {
     @PrimaryKey(autoGenerate = true)
     var id: Long = 0
